+++ bazelrunner/src/main/java/org/jetbrains/bsp/bazel/bazelrunner/utils/Format.java
@@ -1,0 +1,9 @@
+package org.jetbrains.bsp.bazel.bazelrunner.utils;
+
+import java.time.Duration;
+
+public class Format {
+  public static String duration(Duration duration) {
+    return duration.toString().substring(2).replaceAll("(\\d[HMS])(?!$)", "$1 ").toLowerCase();
+  }
+}